Add safety measure when purging devices with data

+++ management/birdhouse_utils.py
@@ -10,6 +10,7 @@ from config import google_geocoder_api_key, bing_geocoder_api_key
 import serial                       # pip install serial
 import serial.tools.list_ports      # pip install pyserial
 import re
+import sys
 # pip install serial geopy pyserial
 
 # These are IDs that are associated with NodeMCU boards
@@ -84,12 +85,54 @@ def update_customer_data(cust_info):
     return cust_info
 
 
-def purge_server_objects_for_device(tbapi, device_name):
+# From https://stackoverflow.com/questions/3041986/apt-command-line-interface-like-yes-no-input
+def query_yes_no(question, default="yes"):
+    """Ask a yes/no question via raw_input() and return their answer.
+
+    "question" is a string that is presented to the user.
+    "default" is the presumed answer if the user just hits <Enter>.
+        It must be "yes" (the default), "no" or None (meaning
+        an answer is required of the user).
+
+    The "answer" return value is True for "yes" or False for "no".
+    """
+    valid = { "yes": True, "y": True, "ye": True, "no": False, "n": False }
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+
+    while True:
+        sys.stdout.write(question + prompt)
+        choice = input().lower()
+        if default is not None and choice == '':
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write("Please respond with 'yes' or 'no' (or 'y' or 'n').\n")
+
+
+def purge_server_objects_for_device(tbapi, device_name, unsafe=False):
     print("Cleaning up server objects for '" + device_name + "'...", end='')
 
     device = tbapi.get_device_by_name(device_name)
     dash = tbapi.get_dashboard_by_name(device_name)
     cust = tbapi.get_customer_by_name(device_name)
+
+    if device is not None and not unsafe:
+        telkeys = tbapi.get_telemetry_keys(device)
+        if len(telkeys) > 0:
+            print()
+            answer = query_yes_no("Device '" + device_name + "' has data.  Are you sure you want to delete it from the server?", default="no")
+            if not answer:
+                print("Aborting!")
+                exit()
+            print("Deleting...", end='')
 
     if dash is not None:
         tbapi.delete_dashboard(tbapi.get_id(dash))
